+++ CHANGELOG.md
@@ -9,10 +9,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
 ### Added
 
 - Add prometheus metrics about number of attempted, succeeded and failed upgrades.
 - Add prometheus metrics about upgrade time of all clusters.
+
+### Added 
+
+- Tests
+
+### Fixed
+
+- Fix out of office time.
 
 ## [0.1.1] - 2021-09-13